moving to lua parse an JavaScript
Moved to lua parse, luajit isn't required anymore.

No need for an external executable, this will lint lua with javascript and potentially do it better.

This also means that the parser now understands 5.3 and 5.2 features.

I've also moved the package to javascript as i userstand it better than coffescript and well it can be minimised.